requirements.txt
# API and data handling
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# LLM model and tokenization
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
